Initial conversion to monorepo and split WebSocket event logic
