+++ .github/workflows/deploy-docs.yml
@@ -25,6 +25,9 @@ jobs:
       - name: Install dependencies
         run: npm ci
 
+      - name: Build packages
+        run: npm run build
+
       - name: Generate TypeDoc documentation
         run: npm run docs
