+++ .prettierignore
@@ -1,3 +1,4 @@
 # Ignore artifacts:
 build
 coverage
+packages/*/dist
