+++ packages/mailpit-api/package.json
@@ -1,0 +1,66 @@
+{
+  "name": "mailpit-api",
+  "version": "2.0.0",
+  "description": "A TypeScript client for interacting with Mailpit's REST API.",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/mpspahr/mailpit-api.git",
+    "directory": "packages/mailpit-api"
+  },
+  "homepage": "https://mpspahr.github.io/mailpit-api/",
+  "bugs": {
+    "url": "https://github.com/mpspahr/mailpit-api/issues"
+  },
+  "publishConfig": {
+    "registry": "https://registry.npmjs.org/",
+    "provenance": true,
+    "access": "public"
+  },
+  "engines": {
+    "node": ">=18.0.0"
+  },
+  "main": "dist/index.cjs",
+  "module": "dist/index.mjs",
+  "exports": {
+    ".": {
+      "import": {
+        "types": "./dist/index.d.mts",
+        "default": "./dist/index.mjs"
+      },
+      "require": {
+        "types": "./dist/index.d.cts",
+        "default": "./dist/index.cjs"
+      }
+    }
+  },
+  "files": [
+    "dist",
+    "README.md",
+    "LICENSE"
+  ],
+  "scripts": {
+    "build": "tsdown src/index.ts --dts --format esm,cjs --outDir dist"
+  },
+  "keywords": [
+    "mailpit",
+    "api",
+    "client",
+    "email",
+    "smtp",
+    "typescript",
+    "test",
+    "playwright",
+    "e2e",
+    "nodejs",
+    "testing",
+    "automation",
+    "integration"
+  ],
+  "author": "Matthew Spahr",
+  "license": "MIT",
+  "typedocOptions": {
+    "entryPoints": [
+      "./src/index.ts"
+    ]
+  }
+}
