+++ tsconfig.json
@@ -2,7 +2,6 @@
   "compilerOptions": {
     "allowJs": true,
     "allowSyntheticDefaultImports": true,
-    "baseUrl": "src",
     "declaration": true,
     "esModuleInterop": true,
     "inlineSourceMap": false,
@@ -13,16 +12,13 @@
     "noFallthroughCasesInSwitch": true,
     "pretty": true,
     "resolveJsonModule": true,
-    "rootDir": ".",
     "skipLibCheck": true,
     "strict": true,
     "traceResolution": false,
     "types": ["node"],
     "module": "esnext",
-    "outDir": "dist",
     "target": "esnext"
   },
   "compileOnSave": false,
-  "exclude": ["node_modules", "dist", "test", "docs", "coverage"],
-  "include": ["src/**/*.ts", "tests/**/*.spec.ts"]
+  "exclude": ["node_modules", "dist", "test", "docs", "coverage"]
 }
